backend/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
